[MODIFY] lotto step3 불필요한 클래스, 메소드 삭제 및 로직 수정 * FixedLottoGenerator, RandomLottoGenerator 삭제하여 로또 생성 관련 메소드 LottoGenerator로 응집 * LottoNumber의 increase 메소드 삭제 * Prize의 valueOf 메소드 로직 스트림으로 수정
