+++ src/main/java/lotto/service/LottoGame.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.service;
+
+import lotto.*;
 
 import java.util.List;
